Add EnduranceTestResult.cs, GeminiModels.cs, SystemMonitor.cs, AIAnalysisService.cs, and HttpClientProvider.cs
