+++ HttpClientProvider.cs
@@ -1,11 +1,13 @@
-﻿using System.Net.Http;
+﻿using System;
+using System.Net.Http;
 
 namespace Endurance_Testing
 {
     public static class HttpClientProvider
     {
-        private static readonly HttpClient _httpClient = new HttpClient();
+        private static readonly Lazy<HttpClient> _httpClientInstance =
+            new Lazy<HttpClient>(() => new HttpClient());
 
-        public static HttpClient Instance => _httpClient;
+        public static HttpClient Instance => _httpClientInstance.Value;
     }
 }
