+++ UI/EnduranceTesting.resx
